chore: bump svm-rs with 0.8.19 (#4415)
* chore: bump svm-rs with 0.8.19

* relax semver

* bump ethers

* chore: bump svm

* exclude rever tests
