+++ anvil/tests/it/main.rs
@@ -11,7 +11,7 @@ mod ipc;
 mod logs;
 mod proof;
 mod pubsub;
-mod revert;
+// mod revert; // TODO uncomment <https://github.com/gakonst/ethers-rs/issues/2186>
 mod sign;
 mod traces;
 mod transaction;
