+++ cli/Cargo.toml
@@ -88,7 +88,7 @@ pretty_assertions = "1.0.0"
 toml = "0.7.0"
 serial_test = "1.0.0"
 criterion = "0.4.0"
-svm = { package = "svm-rs", version = "0.2.16", default-features = false, features = ["rustls"] }
+svm = { package = "svm-rs", version = "0.2", default-features = false, features = ["rustls"] }
 
 [features]
 default = ["rustls"]
