+++ cli/tests/it/svm.rs
@@ -11,7 +11,7 @@ use svm::{self, Platform};
 ///     2. svm updated with all build info
 ///     3. svm bumped in ethers-rs
 ///     4. ethers bumped in foundry + update the `LATEST_SOLC`
-const LATEST_SOLC: Version = Version::new(0, 8, 18);
+const LATEST_SOLC: Version = Version::new(0, 8, 19);
 
 macro_rules! ensure_svm_releases {
     ($($test:ident => $platform:ident),*) => {
